Synchronizing upstream

+++ flink-test-utils-parent/flink-test-utils/src/main/java/org/apache/flink/test/util/TestBaseUtils.java
@@ -23,6 +23,7 @@ import org.apache.flink.api.java.tuple.Tuple;
 import org.apache.flink.configuration.AkkaOptions;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
+import org.apache.flink.configuration.CoreOptions;
 import org.apache.flink.configuration.HighAvailabilityOptions;
 import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.configuration.WebOptions;
@@ -42,6 +43,8 @@ import org.apache.commons.io.IOUtils;
 import org.junit.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import javax.annotation.Nonnull;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
@@ -92,7 +95,7 @@ public class TestBaseUtils extends TestLogger {
 
 	protected static final int MINIMUM_HEAP_SIZE_MB = 192;
 
-	protected static final long TASK_MANAGER_MEMORY_SIZE = 80;
+	protected static final String TASK_MANAGER_MEMORY_SIZE = "80m";
 
 	protected static final long DEFAULT_AKKA_ASK_TIMEOUT = 1000;
 
@@ -101,6 +104,9 @@ public class TestBaseUtils extends TestLogger {
 	public static final FiniteDuration DEFAULT_TIMEOUT = new FiniteDuration(DEFAULT_AKKA_ASK_TIMEOUT, TimeUnit.SECONDS);
 
 	public static final Time DEFAULT_HTTP_TIMEOUT = Time.seconds(10L);
+
+	static final String NEW_CODEBASE = "new";
+	static final String CODEBASE_KEY = "codebase";
 
 	// ------------------------------------------------------------------------
 
@@ -126,7 +132,7 @@ public class TestBaseUtils extends TestLogger {
 		Configuration config = new Configuration();
 
 		config.setInteger(ConfigConstants.LOCAL_NUMBER_TASK_MANAGER, numTaskManagers);
-		config.setInteger(ConfigConstants.TASK_MANAGER_NUM_TASK_SLOTS, taskManagerNumSlots);
+		config.setInteger(TaskManagerOptions.NUM_TASK_SLOTS, taskManagerNumSlots);
 
 		config.setBoolean(ConfigConstants.LOCAL_START_WEBSERVER, startWebserver);
 
@@ -142,22 +148,41 @@ public class TestBaseUtils extends TestLogger {
 		Configuration config,
 		boolean singleActorSystem) throws Exception {
 
-		logDir = File.createTempFile("TestBaseUtils-logdir", null);
-		Assert.assertTrue("Unable to delete temp file", logDir.delete());
-		Assert.assertTrue("Unable to create temp directory", logDir.mkdir());
-		Path logFile = Files.createFile(new File(logDir, "jobmanager.log").toPath());
-		Files.createFile(new File(logDir, "jobmanager.out").toPath());
+		if (!config.contains(WebOptions.LOG_PATH) || !config.containsKey(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY)) {
+			logDir = File.createTempFile("TestBaseUtils-logdir", null);
+			Assert.assertTrue("Unable to delete temp file", logDir.delete());
+			Assert.assertTrue("Unable to create temp directory", logDir.mkdir());
+			Path logFile = Files.createFile(new File(logDir, "jobmanager.log").toPath());
+			Files.createFile(new File(logDir, "jobmanager.out").toPath());
 
-		config.setLong(TaskManagerOptions.MANAGED_MEMORY_SIZE, TASK_MANAGER_MEMORY_SIZE);
-		config.setBoolean(ConfigConstants.FILESYSTEM_DEFAULT_OVERWRITE_KEY, true);
+			if (!config.contains(WebOptions.LOG_PATH)) {
+				config.setString(WebOptions.LOG_PATH, logFile.toString());
+			}
 
-		config.setString(AkkaOptions.ASK_TIMEOUT, DEFAULT_AKKA_ASK_TIMEOUT + "s");
-		config.setString(AkkaOptions.STARTUP_TIMEOUT, DEFAULT_AKKA_STARTUP_TIMEOUT);
+			if (!config.containsKey(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY)) {
+				config.setString(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY, logFile.toString());
+			}
+		}
 
-		config.setInteger(WebOptions.PORT, 8081);
-		config.setString(WebOptions.LOG_PATH, logFile.toString());
+		if (!config.contains(WebOptions.PORT)) {
+			config.setInteger(WebOptions.PORT, 8081);
+		}
 
-		config.setString(ConfigConstants.TASK_MANAGER_LOG_PATH_KEY, logFile.toString());
+		if (!config.contains(AkkaOptions.ASK_TIMEOUT)) {
+			config.setString(AkkaOptions.ASK_TIMEOUT, DEFAULT_AKKA_ASK_TIMEOUT + "s");
+		}
+
+		if (!config.contains(AkkaOptions.STARTUP_TIMEOUT)) {
+			config.setString(AkkaOptions.STARTUP_TIMEOUT, DEFAULT_AKKA_STARTUP_TIMEOUT);
+		}
+
+		if (!config.contains(CoreOptions.FILESYTEM_DEFAULT_OVERRIDE)) {
+			config.setBoolean(CoreOptions.FILESYTEM_DEFAULT_OVERRIDE, true);
+		}
+
+		if (!config.contains(TaskManagerOptions.MANAGED_MEMORY_SIZE)) {
+			config.setString(TaskManagerOptions.MANAGED_MEMORY_SIZE, TASK_MANAGER_MEMORY_SIZE);
+		}
 
 		LocalFlinkMiniCluster cluster =  new LocalFlinkMiniCluster(config, singleActorSystem);
 
@@ -651,6 +676,26 @@ public class TestBaseUtils extends TestLogger {
 		}
 
 		throw new TimeoutException("Could not get HTTP response in time since the service is still unavailable.");
+	}
+
+	@Nonnull
+	public static CodebaseType getCodebaseType() {
+		return Objects.equals(NEW_CODEBASE, System.getProperty(CODEBASE_KEY)) ? CodebaseType.NEW : CodebaseType.LEGACY;
+	}
+
+	public static boolean isNewCodebase() {
+		return CodebaseType.NEW == getCodebaseType();
+	}
+
+	/**
+	 * Type of the mini cluster to start.
+	 *
+	 * @deprecated Will be irrelevant once the legacy mode has been removed.
+	 */
+	@Deprecated
+	public enum CodebaseType {
+		LEGACY,
+		NEW
 	}
 
 	/**
